Major Changes (Read commit description)
V2.1.0 (Yes I am now taking versioning seriously):
Made it so much cleaner and better to understand having each route to its own file, and also added API tracking for more information on who's calling it (Doesn't invade privacy or anything, All it does is track what websites are calling it and which endpoints are being used)
Most of this was done because i got tired of siphoning through my server.js file to find the route I wanted to edit, so this is such a huge code cleanup and a major W for me, 
Also made the deployment test server up-to-date since i was neglecting it for a little bit, but now updates will be consistent on it aswell, tho there wont be any elaborate stuff on it.

Plans for V2.2.0 (Will be updated to 2.2.0 after all the smaller changes listed below are completed):
I plan on making a new card type based on osu lazer's card that appears when a user clicks on their own profile and while making that I will clean out cardRenderer.js.
Cleanup on public folder and making the responsive layout better for mobile users.
